break out pkg/cluster/names to vastly reduce imports of pkg/config, in preparation to make pkg/config version dependent
