+++ pkg/arm/resources.go
@@ -11,6 +11,7 @@ import (
 	"github.com/Azure/go-autorest/autorest/to"
 
 	"github.com/openshift/openshift-azure/pkg/api"
+	"github.com/openshift/openshift-azure/pkg/cluster/names"
 	"github.com/openshift/openshift-azure/pkg/config"
 	"github.com/openshift/openshift-azure/pkg/util/jsonpath"
 	"github.com/openshift/openshift-azure/pkg/util/resourceid"
@@ -468,7 +469,7 @@ func Vmss(cs *api.OpenShiftManagedCluster, app *api.AgentPoolProfile, backupBlob
 			},
 			VirtualMachineProfile: &compute.VirtualMachineScaleSetVMProfile{
 				OsProfile: &compute.VirtualMachineScaleSetOSProfile{
-					ComputerNamePrefix: to.StringPtr(config.GetHostnamePrefix(app, suffix)),
+					ComputerNamePrefix: to.StringPtr(names.GetHostnamePrefix(app, suffix)),
 					AdminUsername:      to.StringPtr(vmssAdminUsername),
 					LinuxConfiguration: &compute.LinuxConfiguration{
 						DisablePasswordAuthentication: to.BoolPtr(true),
@@ -545,7 +546,7 @@ func Vmss(cs *api.OpenShiftManagedCluster, app *api.AgentPoolProfile, backupBlob
 			SinglePlacementGroup: to.BoolPtr(false),
 			Overprovision:        to.BoolPtr(false),
 		},
-		Name:     to.StringPtr(config.GetScalesetName(app, suffix)),
+		Name:     to.StringPtr(names.GetScalesetName(app, suffix)),
 		Type:     to.StringPtr("Microsoft.Compute/virtualMachineScaleSets"),
 		Location: to.StringPtr(cs.Location),
 	}
