+++ pkg/cluster/names/names.go
@@ -1,4 +1,4 @@
-package config
+package names
 
 import (
 	"fmt"
