+++ test/e2e/specs/fakerp/reimagevm.go
@@ -12,7 +12,7 @@ import (
 
 	"k8s.io/apimachinery/pkg/util/wait"
 
-	"github.com/openshift/openshift-azure/pkg/config"
+	"github.com/openshift/openshift-azure/pkg/cluster/names"
 	"github.com/openshift/openshift-azure/pkg/util/resourceid"
 	"github.com/openshift/openshift-azure/test/clients/azure"
 	"github.com/openshift/openshift-azure/test/sanity"
@@ -47,7 +47,7 @@ var _ = Describe("Reimage VM E2E tests [ReimageVM][Fake][LongRunning]", func() {
 		err = azurecli.OpenShiftManagedClustersAdmin.Reimage(context.Background(), os.Getenv("RESOURCEGROUP"), os.Getenv("RESOURCEGROUP"), vm)
 		Expect(err).NotTo(HaveOccurred())
 
-		scaleset, instanceID, err := config.GetScaleSetNameAndInstanceID(vm)
+		scaleset, instanceID, err := names.GetScaleSetNameAndInstanceID(vm)
 		Expect(err).NotTo(HaveOccurred())
 
 		wait.PollImmediate(10*time.Second, 2*time.Minute, func() (bool, error) {
